added new LDA
git-svn-id: https://svn.rice.edu/r/software_0/ObjectQueryModel@3536 522aa1da-077c-40de-a48b-1771163b7009

+++ src/sharedLibraries/headers/LDADocIDAggregate.h
@@ -43,7 +43,7 @@ public:
 
         // the key type must have == and size_t hash () defined
         Lambda <int> getKeyProjection (Handle <LDADocument> aggMe) override {
-                return makeLambda (aggMe, [] (Handle<LDADocument> & aggMe) {return aggMe->getDoc();});
+                return makeLambda (aggMe, [] (Handle<LDADocument> & aggMe) {return (int) aggMe->getDoc();});
         }
 
         // the value type must have + defined
